+++ mapoi_rviz_plugins/src/mapoi_panel.cpp
@@ -185,193 +185,12 @@ void MapoiPanel::onDisable()
   parentWidget()->hide();
 }
 
-void MapoiPanel::MapComboBox()
-{
-  const int index = ui_->MapComboBox->currentIndex();
-  if (index < 0 || index >= static_cast<int>(map_name_list_.size())) {
-    return;
-  }
-  std_msgs::msg::String msg;
-  msg.data = map_name_list_[index];
-  if (mapoi_switch_map_pub_->get_subscription_count() == 0) {
-    RCLCPP_WARN(LOGGER, "mapoi/nav/switch_map has no subscribers; mapoi_nav2_bridge may not be running.");
-  }
-  mapoi_switch_map_pub_->publish(msg);
-  RCLCPP_INFO(LOGGER, "Published map switch request: %s", msg.data.c_str());
-}
-
-void MapoiPanel::Nav2GoalComboBox()
-{
-  goal_combobox_ind_ = ui_->Nav2GoalComboBox->currentIndex();
-  if (goal_combobox_ind_ >= 0 && goal_combobox_ind_ < static_cast<int>(pois_.size())) {
-    highlighted_goal_name_ = pois_[goal_combobox_ind_].name;
-  } else {
-    highlighted_goal_name_.clear();
-  }
-  PublishHighlightPois();
-}
-
-void MapoiPanel::MapoiRouteComboBox()
-{
-  route_combobox_ind_ = ui_->MapoiRouteComboBox->currentIndex();
-  highlighted_route_poi_names_.clear();
-  if (route_combobox_ind_ < 1) {
-    PublishHighlightPois();
-    return;
-  }
-  int route_index = route_combobox_ind_ - 1;
-  if (route_index >= 0 && route_index < static_cast<int>(route_name_list_.size())) {
-    if (get_route_pois_client_->wait_for_service(3s)) {
-      auto request = std::make_shared<mapoi_interfaces::srv::GetRoutePois::Request>();
-      request->route_name = route_name_list_[route_index];
-      auto result = get_route_pois_client_->async_send_request(request);
-      if (rclcpp::spin_until_future_complete(service_node_, result, 5s) == rclcpp::FutureReturnCode::SUCCESS) {
-        auto response = result.get();
-        if (response && response->success) {
-          for (const auto & poi : response->pois_list) {
-            highlighted_route_poi_names_.push_back(poi.name);
-          }
-        } else if (response) {
-          // #342: route が見つからない (typo 等)。highlight は設定しない
-          // (highlighted_route_poi_names_ は上で既に clear 済み)。
-          RCLCPP_ERROR(LOGGER, "get_route_pois failed for '%s': %s",
-                       route_name_list_[route_index].c_str(), response->error_message.c_str());
-        }
-      }
-    } else {
-      RCLCPP_ERROR(LOGGER, "mapoi/get_route_pois service not available after 3s timeout.");
-    }
-  }
-  PublishHighlightPois();
-}
-
-void MapoiPanel::LocalizationButton()
-{
-  // POI 選択ガード: GoalComboBox で選択されていない / 範囲外なら何もしない。
-  // (LocalizationButton は GoalComboBox の POI を流用する仕様、cf. PublishHighlightPois)
-  if (goal_combobox_ind_ < 0 || goal_combobox_ind_ >= static_cast<int>(pois_.size())) {
-    RCLCPP_WARN(LOGGER, "No POI selected for initial pose; pick one in the goal ComboBox first.");
-    return;
-  }
-  // #211: `/initialpose` も `mapoi/initialpose_poi` も直接叩かず、唯一の writer である mapoi_server に
-  // request_initial_pose service 経由で publish を依頼する。これにより transient_local の
-  // per-writer latched cache が単一化され stale POI を構造的に排除する。bridge が POI resolve /
-  // landmark 排他 / `/initialpose` 配信 / retry を一元処理する点は不変 (#209)。
-  if (!request_initial_pose_client_->wait_for_service(3s)) {
-    RCLCPP_ERROR(LOGGER, "mapoi/request_initial_pose service not available after 3s timeout.");
-    return;
-  }
-  auto request = std::make_shared<mapoi_interfaces::srv::RequestInitialPose::Request>();
-  request->map_name = current_map_;
-  request->poi_name = pois_[goal_combobox_ind_].name;
-  auto result = request_initial_pose_client_->async_send_request(request);
-  if (rclcpp::spin_until_future_complete(service_node_, result, 5s) == rclcpp::FutureReturnCode::SUCCESS) {
-    // #299: server は非空 map_name が現在 map と不一致な要求を publish せず success=false で
-    // 返すようになった (panel の current_map_ が stale な窓など)。future 完了 = 成功ではない
-    // ので response を確認し、reject を operator に成功と誤認させない。
-    const auto response = result.get();
-    if (response->success) {
-      RCLCPP_INFO(LOGGER, "Requested initial pose: %s (map: %s).",
-                  request->poi_name.c_str(), current_map_.c_str());
-    } else {
-      RCLCPP_ERROR(LOGGER, "request_initial_pose rejected: %s",
-                   response->error_message.c_str());
-    }
-  } else {
-    RCLCPP_ERROR(LOGGER, "request_initial_pose call failed or timed out.");
-  }
-}
-
-void MapoiPanel::RunGoalButton()
-{
-  if (goal_combobox_ind_ < 0 || goal_combobox_ind_ >= static_cast<int>(pois_.size())) {
-    return;
-  }
-  // Nav2 native `goal_pose` を直接叩かず、`mapoi/nav/goal_pose_poi` 経由で bridge に POI 名を
-  // 渡す (#209 review #3)。bridge が POI resolve → Nav2 action 起動 → `mapoi/nav/status` 配信を
-  // 担当するので、status / cancel / pause / resume の状態管理が WebUI と一貫する。
-  std_msgs::msg::String msg;
-  msg.data = pois_[goal_combobox_ind_].name;
-  mapoi_goal_pose_poi_pub_->publish(msg);
-
-  current_nav_mode_ = "goal";
-  current_nav_target_ = msg.data;
-  ui_->NavStatusLabel->setText(
-      QString::fromStdString("目的地走行中: " + current_nav_target_));
-  RCLCPP_INFO(LOGGER, "Published goal POI request: %s", current_nav_target_.c_str());
-}
-
-void MapoiPanel::RunRouteButton()
-{
-  if (route_combobox_ind_ < 1) {
-    return;
-  }
-  int route_index = route_combobox_ind_ - 1;
-  if (route_index >= static_cast<int>(route_name_list_.size())) {
-    return;
-  }
-  std_msgs::msg::String msg;
-  msg.data = route_name_list_[route_index];
-  mapoi_route_pub_->publish(msg);
-
-  current_nav_mode_ = "route";
-  current_nav_target_ = route_name_list_[route_index];
-  ui_->NavStatusLabel->setText(
-      QString::fromStdString("ルート走行中: " + current_nav_target_));
-  RCLCPP_INFO(LOGGER, "A route was set: %s", msg.data.c_str());
-}
-
-void MapoiPanel::PauseButton()
-{
-  std_msgs::msg::String msg;
-  msg.data = "mapoi_panel";
-  mapoi_pause_pub_->publish(msg);
-  ui_->NavStatusLabel->setText(QString::fromStdString("一時停止中"));
-  RCLCPP_INFO(LOGGER, "Pause requested");
-}
-
-void MapoiPanel::ResumeButton()
-{
-  std_msgs::msg::String msg;
-  msg.data = "mapoi_panel";
-  mapoi_resume_pub_->publish(msg);
-  ui_->NavStatusLabel->setText(QString::fromStdString("再開中..."));
-  RCLCPP_INFO(LOGGER, "Resume requested");
-}
-
-void MapoiPanel::StopButton()
-{
-  std_msgs::msg::String msg;
-  msg.data = "mapoi_panel";
-  mapoi_cancel_pub_->publish(msg);
-
-  current_nav_mode_ = "idle";
-  ui_->NavStatusLabel->setText(QString::fromStdString("走行キャンセル"));
-  RCLCPP_INFO(LOGGER, "The robot was stopped");
-}
+// MapComboBox / Nav2GoalComboBox / MapoiRouteComboBox / LocalizationButton / RunGoalButton /
+// RunRouteButton / PauseButton / ResumeButton / StopButton / PublishHighlightPois は
+// mapoi_panel_nav_control.cpp (#397 step 4) へ移動。
 
 // BackendStatusCallback / LocalizationBackendStatusCallback / UpdateNavButtonsEnabled は
 // mapoi_panel_backend_status.cpp (#397 step 2) へ移動。
-
-void MapoiPanel::PublishHighlightPois()
-{
-  if (mapoi_highlight_goal_pub_) {
-    std_msgs::msg::String goal_msg;
-    goal_msg.data = highlighted_goal_name_;
-    mapoi_highlight_goal_pub_->publish(goal_msg);
-  }
-
-  if (mapoi_highlight_route_pub_) {
-    std_msgs::msg::String route_msg;
-    for (const auto & name : highlighted_route_poi_names_) {
-      if (!route_msg.data.empty()) {
-        route_msg.data += ",";
-      }
-      route_msg.data += name;
-    }
-    mapoi_highlight_route_pub_->publish(route_msg);
-  }
-}
 
 }  // mapoi_rviz_plugins
 CLASS_LOADER_REGISTER_CLASS(mapoi_rviz_plugins::MapoiPanel, rviz_common::Panel)
